@@ -44,12 +44,16 @@ func (d *MongoDBDiscovery) Discover(sessionHandler servicediscovery.ISessionHand
 	clientOptions := options.Client().ApplyURI(fmt.Sprintf("mongodb://%s:%d", sessionHandler.GetHost(), sessionHandler.GetPort()))
 	ctx := context.Background()
 	client, err := mongo.Connect(ctx, clientOptions)
+	defer client.Disconnect(ctx)
 	if err != nil {
-		fmt.Printf("failed to connect to MongoDB server: %v\n", err)
-		return nil, fmt.Errorf("failed to connect to MongoDB server: %v", err)
+		return &MongoDBDiscoveryResult{
+			isDetected:      false,
+			isAuthenticated: true,
+			properties:      nil, // Set properties to nil as it's not used in this case
+		}, nil
 	}
 
-	// Here: we know it is MongoDB, but we don't know if it is authenticated or not
+	// Here: we know it is MongoDB, but we don't know if it's authenticated or not.
 	result := &MongoDBDiscoveryResult{
 		isDetected:      true,
 		isAuthenticated: true,
@@ -63,19 +67,26 @@ func (d *MongoDBDiscovery) Discover(sessionHandler servicediscovery.ISessionHand
 	}
 
 	// Get MongoDB server version
-	serverStatusCmd := bson.D{{"serverStatus", 1}, {"recordStats", 0}}
+	serverStatusCmd := bson.D{{Key: "serverStatus", Value: 1}, {Key: "recordStats", Value: 0}}
 	serverStatusResult := client.Database("admin").RunCommand(ctx, serverStatusCmd)
 	if serverStatusResult.Err() == nil {
 		var resultDoc bson.D
 		err = serverStatusResult.Decode(&resultDoc)
 		if err == nil {
-			version, _ := resultDoc.Map()["version"].(string)
-			host, _ := resultDoc.Map()["host"].(string)
-			result.isAuthenticated = false
-			result.properties = map[string]interface{}{
-				"host":    host,
-				"version": version,
+			bsonBytes, _ := bson.MarshalExtJSON(resultDoc, false, false)
+			var resultDocMap map[string]interface{}
+			err = bson.UnmarshalExtJSON(bsonBytes, false, &resultDocMap)
+			if err == nil {
+				version := resultDocMap["version"].(string)
+				host := resultDocMap["host"].(string)
+				result.properties = map[string]interface{}{
+					"version": version,
+					"host":    host,
+				}
+			} else {
+				fmt.Printf("failed to decode server status result: %v\n", err)
 			}
+			result.isAuthenticated = false
 		} else {
 			fmt.Printf("failed to decode server status result: %v\n", err)
 		}
@@ -83,6 +94,5 @@ func (d *MongoDBDiscovery) Discover(sessionHandler servicediscovery.ISessionHand
 		fmt.Printf("failed to retrieve server status: %v\n", serverStatusResult.Err())
 	}
 
-	client.Disconnect(ctx)
 	return result, nil
 }
